requirements.txt
smplx[all]
autograd
Cython
easydict
graphviz
h5py
lmdb
matplotlib
multiprocess
numpy
opencv-python
pathos
Pillow
tqdm
scikit-image
scikit-learn
scipy
sklearn
chumpy
pytorch-lightning==1.3.5
pyvista
pyrender
seaborn
protobuf~=3.19.0
progress
imutils
pytube
pyyaml
pynvml
opencv-python